feat: RestaurantCards 컴포넌트 TS로 변경, data attribute에 전체 데이터가 아닌 sort, category로
