+++ docs/REFACTOR.md
@@ -1,0 +1,5 @@
+### 1단계 피드백 반영
+
+- [x] 실행 오류...
+- [x] 테스트 코드 분리 및 명세 자세히
+- [x] 웹 컴포넌트 - data-xxx 속성 활용방법 바꾸기 -> 전체 데이터가 아닌 sort 조건, category 필터 조건으로
